fix spacing between sections within Footer
