--- src/components/Footer/elements.js
@@ -1,30 +1,0 @@
-import styled from 'styled-components'
-import { Row } from '../../components/grid'
-
-export const FooterStyled = styled.footer`
-  background: #232323;
-  color: ${props => props.theme.colors.white};
-`
-
-export const FollowUs = styled.section`
-  background: ${props => props.theme.colors.black};
-`
-
-export const Social = styled.ul`
-  display: flex;
-  flex-wrap: wrap;
-
-  @media (max-width: 500px) {
-    li {
-      margin-bottom: 24px;
-    }
-  }
-
-  li:not(:last-child) {
-    margin-right: 24px;
-  }
-`
-
-export const OfficeStyled = styled(Row)`
-  overflow: hidden;
-`
